transcription/pyproject.toml
[project]
name = "transcription"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "grpcio>=1.76.0",
  "livekit>=1.0.17",
  "livekit-api>=1.0.7",
  "openai-whisper>=20250625",
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "python-dotenv>=1.1.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
